fix(interp): plumb Float32 through full method dispatch + regression spec
Mirror W5-I's Value::UInt width-tag fix. Add Value::Float32(f32) so the
Rust seed interpreter preserves IEEE 754 single-precision semantics in
arithmetic, casts, and method dispatch. Before this fix, the canonical
distinguishing test 0.1f32 + 0.2f32 - 0.3f32 returned the f64 error
5.551e-17 instead of the correct f32 result 0.0.

Plumbed across 10 files:
- value.rs: new Value::Float32 variant
- value_impl.rs: as_int/as_float/to_display_string/type_name/value_kind/
  to_debug_string/truthy/matches_type
- value_pointers.rs: Clone + cross-variant PartialEq
- value_bridge.rs: widen-to-f64 at FFI boundary (no f32 bridge tag)
- interpreter/expr/literals.rs: TypedFloat F32 -> Value::Float32 (round
  to f32 at parse time so storage matches IEEE 754 single-precision)
- interpreter/expr/casting.rs: NumericType::F32 -> Float32 cast and
  Float32 input handled in cast_to_numeric / bool / string
- interpreter/expr/ops.rs: new float_kind() and as_f32() helpers;
  Add/Sub/Mul/Div/Mod/Pow/Neg dispatch on f32 vs f64
- interpreter_method/mod.rs: Float32 receiver routes through
  handle_float_methods with Float -> Float32 narrowing on output
- interpreter_helpers/method_dispatch.rs: Float32 chained-method arm;
  Int/Float to_f32 now return Float32 instead of widening
- interpreter_method/primitives.rs: Int/Float to_f32 return Float32

Spec: test/unit/compiler/interpreter/f32_precision_spec.spl — 8 cases
covering 0.1f32+0.2f32-0.3f32 IEEE 754, f32 mantissa boundary at 2^24,
f32 overflow to +inf, f32 div precision, round-trip f64 -> f32 -> f64,
f32 negation, plus f64 no-regression pins.

Empirical pre-fix: '0.00000000000000005551115123125783' (f64 error).
Empirical post-fix: '0' (correct f32 IEEE 754).

Refs feedback_bug_doc_fixes_are_guesses.md (verified bug empirically
before fixing, not from framing).

+++ src/compiler_rust/compiler/src/interpreter_method/primitives.rs
@@ -29,7 +29,7 @@ pub fn handle_int_methods(
         "is_even" => Value::Bool(n % 2 == 0),
         "is_odd" => Value::Bool(n % 2 != 0),
         "to_float" | "to_f64" => Value::Float(n as f64),
-        "to_f32" => Value::Float(n as f32 as f64),
+        "to_f32" => Value::Float32(n as f32),
         "to_i8" | "to_i16" | "to_i32" | "to_i64" | "to_u8" | "to_u16" | "to_u32" | "to_u64" => {
             // Strip "to_" prefix -> numeric type name. Reuse the same cast
             // rules as `expr as <T>` so behavior is consistent.
@@ -267,7 +267,7 @@ pub fn handle_float_methods(
         "is_finite" => Value::Bool(f.is_finite()),
         "to_int" | "truncate" => Value::Int(f.trunc() as i64),
         "to_f64" => Value::Float(f),
-        "to_f32" => Value::Float(f as f32 as f64),
+        "to_f32" => Value::Float32(f as f32),
         "to_i8" | "to_i16" | "to_i32" | "to_i64" | "to_u8" | "to_u16" | "to_u32" | "to_u64" => {
             let tname = &method[3..];
             let nt = NumericType::from_name(tname).expect("handled by match arm above");
